refactor: 조회 메서드 readOnly 적용 및 패키지명 변경
