+++ src/test/java/com/example/barrier_free/BarrierFreeApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.barrierfree;
+package com.example.barrier_free;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
